==> meta/01.runMOFA_meta.Rmd <==
---
title: "Using MOFAcell to perform metaanalysis"
author: "Ricardo Ramirez"
date: "10/31/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r, message=FALSE, warning=FALSE}
library(MOFA2)
library(tidyverse)
library(HDF5Array)
library(ComplexHeatmap)
library(scater)
library(scran)
library(uwot)
library(edgeR)
library(circlize)
library(compositions)
library(MOFAcellulaR)
```

# Import Broad Data

```{r}
# Defining the patient annotations -----------------------------------------------------------------
broad_meta <- read_csv("./data_HCMDCM_Nature/meta_data.csv")[,-1] %>%
  dplyr::mutate("group" = "Chaffin2022")
# Importing coldata of the matrices
coldat_broad <- read_csv("./data_HCMDCM_Nature/pb_coldata.csv",
                   show_col_types = FALSE)[,-1] %>%
  column_to_rownames("colname") %>%
  dplyr::rename(ncells = "counts")
# Read normalized data
broad_dat <- readRDS("./data_HCMDCM_Nature/pb_red.rds")

broad_meta %>% dplyr::group_by(disease) %>% summarise(n())
```

# Import Science Data

```{r}
# Defining the patient annotations -----------------------------------------------------------------
hub_meta <- read_csv("./data_DCMACM_Science/meta_data.csv")[,-1] %>%
    dplyr::mutate("group" = "Reichart2022")

hub_meta %>% dplyr::group_by(disease) %>% summarise(n())

# Disease DF
disease_dict <- hub_meta$disease %>% 
  unique() %>% 
  enframe(value = "disease", 
          name = "dis_code")

disease_dict$dis_code <- c("NF", "DCM", 
                           "ARVC", "NCC")

hub_meta <- left_join(hub_meta, disease_dict, by = "disease")
hub_meta$disease <- hub_meta$dis_code

# Importing coldata of the matrices
coldat_hub <- read_csv("./data_DCMACM_Science/pb_coldata.csv",
                   show_col_types = FALSE)[,-1] %>%
  column_to_rownames("colname") %>%
  dplyr::rename(ncells = "counts")

# Read normalized data w/ prior genes from broad
#hub_dat <- readRDS("./data_DCMACM_Science/pb_red_wprior.rds")
hub_dat <- readRDS("./data_DCMACM_Science/pb_red.rds")

```

# Per gene take the variable genes that occurr in both

```{r}
#Hubner genes
hub_features <- hub_dat %>%
  dplyr::select(view, feature) %>%
  unique() %>%
  mutate(study = "Reichart2022")

broad_features <- broad_dat %>%
  dplyr::select(view, feature) %>%
  unique() %>%
  mutate(study = "Chaffin2022")

all_features <- bind_rows(hub_features, broad_features) %>% 
  group_by(view, feature) %>% 
  summarize(n_times = n()) %>% 
  dplyr::filter(n_times == 2)

all_features %>%
  summarize(n_genes = n())
```

# Filter Broad data to contain only Science genes

```{r}
broad_dat <- broad_dat %>%
  dplyr::filter(feature %in% unique(all_features$feature))

hub_dat <- hub_dat %>%
  dplyr::filter(feature %in% unique(all_features$feature))
```

# Generating the meta analysis

```{r}
broad_dat <- broad_dat %>%
  mutate(group = "Chaffin2022")

hub_dat <- hub_dat %>%
  mutate(group = "Reichart2022")
```

```{r}
all_dat <- dplyr::bind_rows(broad_dat, hub_dat)

write_csv(all_dat, "./results/META/alldat.csv")
```

```{r}
MOFAobject <- create_mofa(all_dat)

data_opts <- get_default_data_options(MOFAobject)

data_opts$scale_groups <- TRUE

model_opts <- get_default_model_options(MOFAobject)

model_opts$spikeslab_weights <- FALSE

model_opts$num_factors <- 6

train_opts <- get_default_training_options(MOFAobject)

# Prepare MOFA model:
MOFAobject <- prepare_mofa(
  object = MOFAobject,
  data_options = data_opts,
  model_options = model_opts,
  training_options = train_opts
)

# Train model:
outfile <- file.path("./results/META/MOFA_mcell/", "model.hdf5")

model <- run_mofa(MOFAobject, outfile)
```

```{r}
all_meta <- dplyr::bind_rows(broad_meta, hub_meta) %>%
  dplyr::rename("sample" = "donor_id")

write_csv(all_meta, "./results/META/allmeta.csv")
```

# Downstream analysis

## Association of factors to covariates

We will associate clinical covariates to factor scores. For each factor score, we will perform an analysis of variance to clinical covariates of interest. 

In this specific case we will identify which factors associate with the histological description

## Association of factors to covariates

We will associate clinical covariates to factor scores. For each factor score, we will perform an analysis of variance to clinical covariates of interest. 

In this specific case we will identify which factors associate with the histological description

```{r}
all_meta <- all_meta %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "no", "yes"))

expl_var_pgroup <- MOFAcellulaR::get_associations(model = model,
                                       metadata = all_meta,
                                        sample_id_column = "sample",
                                        test_variable = "disease",
                                        test_type = "categorical",
                                        categorical_type = "non-parametric",
                                        group = TRUE)
expl_var_pgroup

expl_var_pgroup %>%
  dplyr::filter(adj_pvalue <= 0.05)
```

```{r}
stat.test <- MOFAcellulaR::get_tidy_factors(model = model, meta = all_meta, 
                                            group = TRUE, factor = "all") %>%
    dplyr::filter(Factor %in% c("Factor1")) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "NF", "HF")) %>%
  ggpubr::compare_means(formula = value ~ HF, 
                        group.by = "group")

print(stat.test)
```

```{r}
stat.test <- MOFAcellulaR::get_tidy_factors(model = model, meta = all_meta, 
                                            group = TRUE, factor = "all") %>%
    dplyr::filter(Factor %in% c("Factor2")) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "NF", "HF")) %>%
  ggpubr::compare_means(formula = value ~ HF, 
                        group.by = "group")

print(stat.test)
```

```{r}
stat.test <- MOFAcellulaR::get_tidy_factors(model = model, meta = all_meta, 
                                            group = TRUE, factor = "all") %>%
    dplyr::filter(Factor %in% c("Factor4")) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "NF", "HF")) %>%
  ggpubr::compare_means(formula = value ~ HF, 
                        group.by = "group")

print(stat.test)
```


```{r}
pdf("./results/META/FactorScores_HFbox.pdf", height = 2.5, width = 2.5)

plt_box <- MOFAcellulaR::get_tidy_factors(model = model, meta = all_meta, 
                                            group = TRUE, factor = "all") %>%
    dplyr::filter(Factor %in% c("Factor1")) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "NF", "HF")) %>%
  ggplot(aes(x = group, y = value, fill = HF)) +
  geom_boxplot() +
  theme_classic() +
  scale_fill_manual(values = c("black", "grey")) +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 12),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust =0.5)) +
  ylab("Factor 1 score") +
  xlab("Study")

plot(plt_box)

dev.off()
```

```{r}
pdf("./results/META/FactorScores_HFbox_B.pdf", height = 2.5, width = 2.5)

plt_box <- MOFAcellulaR::get_tidy_factors(model = model, meta = all_meta, 
                                            group = TRUE, factor = "all") %>%
  dplyr::filter(Factor %in% c("Factor2")) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "NF", "HF")) %>%
  ggplot(aes(x = group, y = value, fill = HF)) +
  geom_boxplot() +
  theme_classic() +
  scale_fill_manual(values = c("black", "grey")) +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 12),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust =0.5)) +
  ylab("Factor 2 score") +
  xlab("Study")

plot(plt_box)

dev.off()
```

```{r}
MOFAcellulaR::get_tidy_factors(model = model, meta = all_meta, 
                                            group = TRUE, factor = "all") %>%
  dplyr::filter(Factor %in% c("Factor4")) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "NF", "HF")) %>%
  ggplot(aes(x = group, y = value, fill = HF)) +
  geom_boxplot() +
  theme_classic() +
  scale_fill_manual(values = c("black", "grey")) +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 12),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust =0.5)) +
  ylab("Factor 4 score") +
  xlab("Study")
```

```{r, fig.height=3, fig.width=4.5}
score_plt <- MOFAcellulaR::get_tidy_factors(model = model, meta = all_meta, 
                                            group = TRUE, factor = "all") %>%
  dplyr::filter(Factor %in% c("Factor1", "Factor2")) %>%
  pivot_wider(names_from = Factor, values_from = value) %>%
  ggplot(., aes(x = Factor1,
                y = Factor2,
                color = disease,
                shape = group)) +
  geom_point(size = 2.5) +
  theme_classic() +
  theme(axis.text = element_text(size =12),
        axis.title = element_text(size =12)) +
  scale_color_manual(values = c("lightblue", "#3CB371", "#4169E1", "pink", "#FF6666"))

plot(score_plt)

pdf("./results/META/FactorScoresScatter.pdf",height = 3,width = 4.5)

plot(score_plt)

dev.off()

```

```{r}
expl_var_sex <- MOFAcellulaR::get_associations(model = model,
                                       metadata = all_meta,
                                        sample_id_column = "sample",
                                        test_variable = "sex",
                                        test_type = "categorical",
                                        categorical_type = "non-parametric",
                                        group = TRUE)
expl_var_sex
```

```{r}
expl_var_study <- MOFAcellulaR::get_associations(model = model,
                                       metadata = all_meta,
                                        sample_id_column = "sample",
                                        test_variable = "group",
                                        test_type = "categorical",
                                        categorical_type = "non-parametric",
                                        group = TRUE)
expl_var_study
```

## Explained variance per factor and summary heatmap

We recover the explained variance per view per factor to evaluate if coordinated gene expression processes are coordinated between all the cell types or exclusive to only a group of them

```{r}
assoc_list <- list(HF = expl_var_pgroup, 
                   study = expl_var_study,
                   sex = expl_var_sex)

col_list <- list(disease = c("ARVC" = "lightblue",
                                                    "NF" = "#FF6666", 
                                                    "HCM" = "#4169E1",
                                                    "DCM"= "#3CB371",
                                                    "NCC" = "pink"),
                                  group = c("Chaffin2022" = "black",
                                            "Reichart2022" = "darkgrey"))

scores_hmap <- MOFAcellulaR::plot_MOFA_hmap(model = model,
                group = TRUE,
                metadata = all_meta,
                sample_id_column = "sample",
                sample_anns = c("disease", "group"),
                assoc_list = assoc_list,
               col_rows = col_list)

pdf("./results/META/MOFA_mcell/MOFAcell_summ.pdf", height = 9, width = 4.5)

draw(scores_hmap)

dev.off()

```

## Paper stats

### Recovered mean across all samples

#Group 0 = Broad
```{r}
model@cache$variance_explained$r2_total$Chaffin2022
model@cache$variance_explained$r2_total$Chaffin2022 %>% mean()
```
#Group 1 = Hubner
```{r}
model@cache$variance_explained$r2_total$Reichart2022
model@cache$variance_explained$r2_total$Reichart2022 %>% mean()
```

### Factors associated with patient group
```{r}
useful_factors <- expl_var_pgroup %>%
  dplyr::filter(adj_pvalue < 0.05) %>%
  pull(Factor)

print(useful_factors)

model@cache$variance_explained$r2_per_factor$Chaffin2022[useful_factors,,drop = F] %>%
  colSums()

model@cache$variance_explained$r2_per_factor$Chaffin2022[useful_factors,,drop = F] %>%
  colSums() %>%
  mean()
```

```{r}
model@cache$variance_explained$r2_per_factor$Reichart2022[useful_factors,,drop = F] %>%
  colSums()

model@cache$variance_explained$r2_per_factor$Reichart2022[useful_factors,,drop = F] %>%
  colSums() %>%
  mean()
```
