==> meta/02.deconvoluteReHeaT.Rmd <==
---
title: "Deconvolute ReHeaT"
author: "Ricardo Ramirez"
date: "11/1/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

## MOFA cell for deconvolution

Here we demonstrate that MOFA cell loadings can be used to deconvolute disease signals from ReHeaT

```{r}
library(tidyverse)
library(MOFA2)
source("./MOFAcell/factorutils.R")
```

Importing the results from the meta-model

```{r}
model_outfile <- "./results/META/MOFA_mcell/model.hdf5"
model <- MOFA2::load_model(model_outfile)
pb_data <- read_csv("./results/META/alldat.csv")
meta <- read_csv("./results/META/allmeta.csv")
```

Getting factor scores and loadings (Factor 1)

```{r}
# Get factor scorees
factor_scores <- get_fscores(model = model,
                               meta = meta,
                               factor = "Factor1", group = T)
# Get loadings
factor_loadings <- get_floadings(model = model,
                                   factor = "Factor1")

```

Import ReHeaT studies, scale normalized expression and calculate weighted means

```{r}
# Calculate disease scores for all samples ----------------------------------
reheat <- readRDS("./data_reheat/METAheart.rds")

dr_res <- map(reheat, function(x) {
  
  mat <- x$GEX %>%
    t() %>%
    scale() %>%
    t()
  
  dR_run <- decoupleR::run_wmean(mat = mat, 
                                 network = factor_loadings,
                                 .source = ctype, 
                                 .target = feature,
                                 .mor = value) %>%
    dplyr::filter(statistic == "norm_wmean")
  
  dR_run <- dR_run  %>%
    left_join(x$TARGETS %>%
                dplyr::select(Sample, HeartFailure),
              by = c("condition" = "Sample"))
  
}) %>%
  enframe() %>%
  unnest()
```

## Make matrix with all scores

```{r}
sample_info <- dr_res %>%
  dplyr::select(condition, HeartFailure, name) %>%
  unique() %>%
  dplyr::mutate(sample_id = paste0(name, "_", condition)) %>%
  column_to_rownames("sample_id")

ds_mat <- dr_res %>%
  dplyr::mutate(sample_id = paste0(name, "_", condition)) %>%
  dplyr::select(sample_id, source, score) %>%
  pivot_wider(names_from = source, values_from = score) %>%
  column_to_rownames("sample_id")

ds_mat <- as.matrix(ds_mat) %>%
  scale()

saveRDS(ds_mat, "./results/META/ds_mofacell_mat.rds")
```

```{r}
ht_opt$ROW_ANNO_PADDING <- unit(2.5, "mm")
ht_opt$COLUMN_ANNO_PADDING <- unit(2.5, "mm")

# Association colors

col_fun_scores <- circlize::colorRamp2(seq(-5, 5, length = 30), hcl.colors(30,"Green-Brown",rev = T))

row_ha <- rowAnnotation(HF = sample_info[rownames(ds_mat),"HeartFailure"],
                       col = list(HF = c("no" = "grey", 
                                                    "yes"= "black")),
                       gap = unit(2.5, "mm"),
                       border = TRUE)

scores_hmap <- Heatmap(ds_mat[, sort(colnames(ds_mat))], 
        name = "cell type \ndisease", 
        right_annotation = row_ha,
        cluster_columns = FALSE, 
        show_row_dend = FALSE,
        show_row_names = FALSE,
        border = TRUE,
        gap = unit(2.5, "mm"),
        col = col_fun_scores)

pdf("./results/META/reheat_ct_scores.pdf", height = 4, width = 2.3)

draw(scores_hmap)

dev.off()

```

# Test the difference between cell-type disease scores across cell-types 

```{r}
ct_ds <- dr_res %>%
  group_by(name, source) %>%
  dplyr::mutate(scale_score = scale(score)[,1]) %>%
  dplyr::mutate(HeartFailure = factor(HeartFailure, 
                                      levels = c("yes", "no"))) %>%
  nest() %>%
  mutate(ds_t = map(data, function(dat) {
    
    mean_scores <- dat %>%
      group_by(HeartFailure) %>%
      summarize(mean_score = mean(scale_score)) %>%
      pivot_wider(names_from = HeartFailure,values_from = mean_score)
    
    # Sign of the difference between the means
    sign_diff <-  mean_scores[,"yes"] - mean_scores[,"no"]
    
    wilcox.test(scale_score ~ HeartFailure, data = dat, alternative = "less") %>%
      broom::tidy() %>%
      mutate(mean_diff = sign(sign_diff[[1]]))
    
  }))

ct_ds_stat <- ct_ds %>%
  dplyr::select(name, ds_t) %>%
  unnest() %>%
  group_by(name) %>%
  dplyr::mutate(p_adj = p.adjust(p.value)) %>%
  dplyr::select(source, name, mean_diff, p.value, p_adj) %>%
  dplyr::mutate(significance = ifelse(p_adj <= 0.1, "*", "")) %>%
  dplyr::mutate(significance = ifelse(p_adj <= 0.05, "**", significance)) %>%
  dplyr::mutate(mean_diff = factor(mean_diff))

saveRDS(ct_ds_stat, "./results/META/mofacell_ds_hftest.rds")
```

# Show the example of one study

```{r, fig.height=3.5, fig.width= 2.5}
example_violin <- ct_ds %>%
  dplyr::filter(name == "Liu15_M") %>%
  dplyr::select(name, source, data) %>%
  unnest() %>%
  dplyr::filter(source != "Lymphoid") %>%
  dplyr::mutate(disease = ifelse(HeartFailure == "yes", "HeartFailure", "NonFailing")) %>%
  ggplot(., aes(x  = disease, y = score)) +
  geom_violin() +
  theme_classic2() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) +
  facet_wrap(.~source, nrow = 3) +
  ylab("Factor 1 program") +
  xlab("") +
  ggtitle("Liu15_M")

plot(example_violin)

pdf("./results/META/bulk_deconv_example.pdf", height = 4, width = 2.3)
plot(example_violin)
dev.off()
```

```{r, fig.height=3.7, fig.width= 3.5}
# Here we are using the t-values
reheat_stats_plt <- ggplot(ct_ds_stat, aes(y = name, 
                       x  = source, 
                       fill = mean_diff,
                       label = significance)) +
  geom_tile() +
  geom_text() +
  theme(axis.text = element_text(size =12),
    axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) +
  ylab("") +
  xlab("") +
  coord_equal() +
  scale_fill_manual(values = c("darkred","darkblue"))

plot(reheat_stats_plt)

pdf("./results/META/reheat_stats_plt.pdf", height = 4, width = 4)
plot(reheat_stats_plt)
dev.off()
```
