==> meta/04.functionalenrichment.Rmd <==
---
title: "Characterize meta-analysis"
author: "Ricardo Ramirez"
date: "2023-01-10"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(tidyverse)
library(MOFA2)
library(decoupleR)
source("./MOFAcell/factorutils.R")
```

Here we use progeny to infer activities from the gene loadings of the meta-model of Reichart2022 and Chaffin2022

## What are the genes that we prioritize in the meta analysis?

```{r}
model_outfile <- "./results/META/MOFA_mcell/model.hdf5"
model <- MOFA2::load_model(model_outfile)
pb_data <- read_csv("./results/META/alldat.csv")
meta <- read_csv("./results/META/allmeta.csv")
```

```{r}
factor_loadings <- get_floadings(model = model,
                                   factor = "Factor1") %>%
  dplyr::filter(abs(value)>=0.1) %>%
  pivot_wider(names_from = ctype, values_from = value, values_fill = 0) %>%
  column_to_rownames("feature") %>%
  as.matrix()

get_floadings(model = model,
              factor = "Factor1") %>%
  dplyr::filter(abs(value)>=0.1) %>% 
  write_csv("./results/META/MOFAcell_metaloadings.csv")
```

## Perform estimation of pathway activities using PROGENy

```{r}
net <- get_progeny(organism = 'human', top = 500)
```

```{r}

col_fun_scores <- circlize::colorRamp2(seq(-5, 5, length = 50), hcl.colors(50,"Blue-Red 2",rev = F))

ct_acts <- run_wmean(mat=factor_loadings, net=net, .source='source', .target='target',
                  .mor='weight', times = 1000, minsize = 5)

ct_acts_mat <- ct_acts %>%
  dplyr::filter(statistic == "norm_wmean") %>%
  dplyr::mutate(score = ifelse(score <= -5, -5, score)) %>%
  dplyr::select(source, condition, score) %>%
  pivot_wider(values_from = score, names_from = condition) %>%
  column_to_rownames("source") %>%
  as.matrix()

hmap <- ComplexHeatmap::Heatmap(t(ct_acts_mat), 
                        name = "progeny", 
                        col = col_fun_scores, 
                        show_row_dend = FALSE,
                        show_column_dend = FALSE)

pdf("./results/META/MOFA_mcell/progeny_res.pdf", height = 2.7, width = 5.5)
draw(hmap)
dev.off()

draw(hmap)
```
