==> meta/05.patientmaps.Rmd <==
---
title: "Patient maps"
author: "Ricardo Ramirez"
date: "2023-04-17"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=F, warning=F}
library(MOFAcellulaR)
library(MOFA2)
library(tidyverse)
library(decoupleR)
library(ranger)
library(ROCR)
library(uwot)
library(cowplot)
set.seed(29)
```

# New patient maps

Here, we will demonstrate that the factors learned from a dataset can be used to project new data and useful to classify unseen data

# Reichart data
### Model
```{r}
reichart_model <- file.path("./results/DCM_hubner/MOFA_mcell/", "DCM_hubner.hdf5") %>% 
  MOFA2::load_model()
```
### Annotations and pseudobulk data
```{r}
# Defining the patient annotations -----------------------------------------------------------------
reichart_sample_meta <- read_csv("./data_DCMACM_Science/meta_data.csv")[,-1]
# Disease DF
reichart_disease_dict <- reichart_sample_meta$disease %>% 
  unique() %>% 
  enframe(value = "disease", 
          name = "dis_code")

reichart_disease_dict$dis_code <- c("NF", "DCM", 
                           "ARVC", "NCC")

reichart_sample_meta <- reichart_sample_meta %>%
  left_join(reichart_disease_dict, by = "disease")

# Importing pb data
reichart_pb_data <- readRDS("./data_DCMACM_Science/pb_red_mat.rds")

# Importing coldata of the matrices
reichart_coldat <- read_csv("./data_DCMACM_Science/pb_coldata.csv",
                   show_col_types = FALSE)[,-1] %>%
  dplyr::mutate(cell_type = strsplit(colname,"_") %>%
                  map_chr(., ~ .x %>% last())) %>%
  column_to_rownames("colname") %>%
  dplyr::rename(ncells = "counts") %>%
  dplyr::rename("cell_counts" = "ncells") 

reichart_sample_meta <- reichart_sample_meta %>%
  dplyr::select(-disease) %>%
  dplyr::rename("sample" = "donor_id",
                "disease" = "dis_code") %>%
  dplyr::mutate(study = "Reichart2022")

reichart_pb_data <- reichart_pb_data[, rownames(reichart_coldat)]
```
### Do light MOFAcellulaR processing
```{r}
# Define views of interest in ct
exclude_ct <- c("Mast", "Neuronal", "prolif", "Adipo", "none")

cts <- reichart_coldat$cell_type %>%
  unique()

cts <- cts[! cts %in% exclude_ct]

reichart_multiview_dat <- MOFAcellulaR::create_init_exp(counts = reichart_pb_data,  
                                               coldata = reichart_coldat) %>%
  MOFAcellulaR::filt_profiles(pb_dat = .,
                          cts = cts,
                          ncells = 25,
                          counts_col = "cell_counts", # This refers to the column name in testcoldata where the number of cells per profile was stored
                          ct_col = "cell_type") %>%
  MOFAcellulaR::tmm_trns(pb_dat_list = .,
                         scale_factor = 1000000) %>%
  MOFAcellulaR::center_views(pb_dat_list = .) %>%
  MOFAcellulaR::pb_dat2MOFA(pb_dat_list = .)
```
### Test effective projection
```{r}
test_projection <- project_data(model = reichart_model, test_data = reichart_multiview_dat)

original_projection <- MOFA2::get_factors(reichart_model)[[1]]

cor(test_projection, original_projection) %>% diag() %>% mean()
```
# Chaffin data
## Model
```{r}
chaffin_model <- file.path("./results/DCM_broad/MOFA_mcell/", "DCMBroad_model.hdf5") %>% 
  MOFA2::load_model()
```
## Expression data
```{r}
chaffin_sample_meta <- read_csv("./data_HCMDCM_Nature/meta_data.csv")[,-1]
# Importing pb data
chaffin_pb_data <- read_csv("./data_HCMDCM_Nature/pb_data.csv",
                    show_col_types = FALSE)

colnames(chaffin_pb_data)[1] <- "sample_id"

chaffin_pb_data <- chaffin_pb_data %>%
  column_to_rownames("sample_id") %>%
  as.matrix() %>%
  t()

# Importing coldata of the matrices
chaffin_coldat <- read_csv("./data_HCMDCM_Nature/pb_coldata.csv",
                   show_col_types = FALSE)[,-1] %>%
  column_to_rownames("colname") %>%
  dplyr::rename(ncells = "counts") %>%
  dplyr::rename("cell_counts" = "ncells")

chaffin_sample_meta <- chaffin_sample_meta %>%
  dplyr::rename("sample" = "donor_id") %>%
  dplyr::mutate(study = "Chaffin2022")
```
### Do light MOFAcellulaR processing
```{r}
exclude_ct <- c("Mast", "Neuronal", "prolif", "Adipo", "none")

cts <- chaffin_coldat$cell_type %>%
  unique()

cts <- cts[! cts %in% exclude_ct]


chaffin_multiview_dat <- MOFAcellulaR::create_init_exp(counts = chaffin_pb_data,  
                                               coldata = chaffin_coldat) %>%
  MOFAcellulaR::filt_profiles(pb_dat = .,
                          cts = cts,
                          ncells = 25,
                          counts_col = "cell_counts", # This refers to the column name in testcoldata where the number of cells per profile was stored
                          ct_col = "cell_type") %>%
  MOFAcellulaR::tmm_trns(pb_dat_list = .,
                         scale_factor = 1000000) %>%
  MOFAcellulaR::center_views(pb_dat_list = .) %>%
  MOFAcellulaR::pb_dat2MOFA(pb_dat_list = .)
```
### Test effective projection
```{r}
test_projection <- project_data(model = chaffin_model, test_data = chaffin_multiview_dat)

original_projection <- MOFA2::get_factors(chaffin_model)[[1]]

cor(test_projection, original_projection) %>% diag() %>% mean()
```

### Make useful joint meta-data

```{r}
all_meta <- bind_rows(chaffin_sample_meta[,c("sample", "disease", "study")],
          reichart_sample_meta[,c("sample", "disease", "study")])
```


# New patient map

The idea is that given an established manifold of multicellular programs that describes the variability of a collection of patients, it is possible to project new patients to that reference.

Some notes about this:

- This is relative to the variability captured by the original collection of patients
- This is limited to the variability captured by the new collection of patients

Here we are going to build pairwise classifiers of HF. For two given studies A and B, we will train a random forest on the multicellular programs of A and B, separately and then we will test how this model predicts HF on unseen projected data.

### Training Chaffin classifier
```{r}
chaffin_RF_dat <- MOFA2::get_factors(chaffin_model)[[1]] %>%
  as.data.frame() %>%
  tibble::rownames_to_column("sample") %>%
  left_join(chaffin_sample_meta[, c("sample", "disease")], by = "sample") %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "no", "yes")) %>%
  dplyr::mutate(HF = factor(HF)) %>%
  dplyr::select(-sample) %>%
  dplyr::select(-disease)

#Compute weights to balance the RF
w <- 1/table(chaffin_RF_dat$HF)
w <- w/sum(w)
weights <- rep(0, nrow(chaffin_RF_dat))
weights[chaffin_RF_dat$HF == "no"] <- w["no"]
weights[chaffin_RF_dat$HF == "yes"] <- w["yes"]

chaffin_RF <- ranger(HF ~ ., data = chaffin_RF_dat, case.weights = weights, probability = TRUE, importance = "permutation")
chaffin_RF

ranger::importance(chaffin_RF)
```

### Projecting Reichart into Chaffin
```{r}
reichart_projection <- MOFAcellulaR::project_data(model = chaffin_model,
                                                  test_data = reichart_multiview_dat) %>%
  as.data.frame() %>%
  tibble::rownames_to_column("sample") %>%
  left_join(reichart_sample_meta[, c("sample", "disease")], by = "sample") %>%
  dplyr::select(-sample) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "no", "yes")) %>%
  dplyr::mutate(HF = factor(HF)) %>%
  dplyr::select(-disease)

real_reichart_labels <- reichart_projection$HF

pred_reichart_labels <- stats::predict(chaffin_RF, reichart_projection %>% dplyr::select(-HF))

pred_reichart_labels <- pred_reichart_labels$predictions

# Final evaluation
reichart_eval <- tibble(real_reichart_labels, prob_healthy = pred_reichart_labels[,"no"]) %>%
  arrange(desc(prob_healthy)) %>%
  dplyr::mutate(real_reichart_labels = ifelse(real_reichart_labels == "no", 1, 0))

PR_object <- prediction(reichart_eval$prob_healthy, reichart_eval$real_reichart_labels) #Evaluate classification
reichart_pr <- ROCR::performance(PR_object, measure = "aucpr")

reichart_pr@y.name
reichart_pr@y.values
```

### Projection plots - Chaffin

Basically we will put the original an projected data - Run an UMAP first with the original data - then with projected

```{r}
chaffin_factor_space <- rbind(MOFA2::get_factors(chaffin_model)[[1]],
                                  MOFAcellulaR::project_data(model = chaffin_model,
                                                  test_data = reichart_multiview_dat))


# UMAP of factor
chaffin_umap <- uwot::umap(chaffin_factor_space, n_neighbors = 7)
colnames(chaffin_umap) <- c("UMAP_1", "UMAP_2")

chaffin_umap <- chaffin_umap %>%
  as.data.frame() %>%
  rownames_to_column("sample") %>%
  left_join(all_meta, by = "sample") %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "no", "yes"))

# First plot Chaffin space
chaffin_space <- chaffin_umap %>%
  dplyr::filter(study == "Chaffin2022") %>%
  ggplot(., aes(x = UMAP_1,
                y = UMAP_2,
                color = HF)) +
  geom_point(size = 2, stroke = 1) +
  theme_classic() +
  theme(axis.text = element_text(size =12)) +
  scale_color_manual(values = c("grey", "black")) +
  xlim(c((chaffin_umap$UMAP_1 %>% min()) - 1,
         (chaffin_umap$UMAP_1 %>% max()) + 1)) +
  ylim(c((chaffin_umap$UMAP_2 %>% min()) - 1,
         (chaffin_umap$UMAP_2 %>% max()) + 1))

reichart_proj <- chaffin_umap %>%
  #plyr::filter(study == "Chaffin2022") %>%
  ggplot(., aes(x = UMAP_1,
                y = UMAP_2,
                color = HF,
                shape = study)) +
  geom_point(size = 2, stroke = 1) +
  theme_classic() +
  theme(axis.text = element_text(size =12)) +
  scale_color_manual(values = c("grey", "black")) +
  xlim(c((chaffin_umap$UMAP_1 %>% min()) - 1,
         (chaffin_umap$UMAP_1 %>% max()) + 1)) +
  ylim(c((chaffin_umap$UMAP_2 %>% min()) - 1,
         (chaffin_umap$UMAP_2 %>% max()) + 1))

panel_chaffin <- cowplot::plot_grid(chaffin_space, reichart_proj, align = "hv", ncol = 1)

pdf("./results/META/chaffin_UMAP.pdf", height = 4.5, width = 3.8)
plot(panel_chaffin)
dev.off()
```

### Training Reichart classifier

```{r}
reichart_RF_dat <- MOFA2::get_factors(reichart_model)[[1]] %>%
  as.data.frame() %>%
  tibble::rownames_to_column("sample") %>%
  left_join(reichart_sample_meta[, c("sample", "disease")], by = "sample") %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "no", "yes")) %>%
  dplyr::mutate(HF = factor(HF)) %>%
  dplyr::select(-sample) %>%
  dplyr::select(-disease)

#Compute weights to balance the RF
w <- 1/table(reichart_RF_dat$HF)
w <- w/sum(w)
weights <- rep(0, nrow(reichart_RF_dat))
weights[reichart_RF_dat$HF == "no"] <- w["no"]
weights[reichart_RF_dat$HF == "yes"] <- w["yes"]

reichart_RF <- ranger(HF ~ ., data = reichart_RF_dat, case.weights = weights, 
                      probability = TRUE, importance = "permutation")
reichart_RF

ranger::importance(reichart_RF)

```

### Projecting Chaffin into Reichart

```{r}
chaffin_projection <- MOFAcellulaR::project_data(model = reichart_model,
                                                test_data = chaffin_multiview_dat) %>%
  as.data.frame() %>%
  tibble::rownames_to_column("sample") %>%
  left_join(chaffin_sample_meta[, c("sample", "disease")], by = "sample") %>%
  dplyr::select(-sample) %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "no", "yes")) %>%
  dplyr::mutate(HF = factor(HF)) %>%
  dplyr::select(-disease)

real_chaffin_labels <- chaffin_projection$HF
pred_chaffin_labels <- stats::predict(reichart_RF, chaffin_projection %>% dplyr::select(-HF))
pred_chaffin_labels <- pred_chaffin_labels$predictions

# Final evaluation
chaffin_eval <- tibble(real_chaffin_labels, prob_healthy = pred_chaffin_labels[,"no"]) %>%
  arrange(desc(prob_healthy)) %>%
  dplyr::mutate(real_chaffin_labels = ifelse(real_chaffin_labels == "no", 1, 0))

PR_object <- prediction(chaffin_eval$prob_healthy, chaffin_eval$real_chaffin_labels) #Evaluate classification
chaffin_pr <- ROCR::performance(PR_object, measure = "aucpr")

chaffin_pr@y.name
chaffin_pr@y.values
```

### Projection plots - Reichart

Basically we will put the original an projected data - Run an UMAP first with the original data - then with projected

```{r}
reichart_factor_space <- rbind(MOFA2::get_factors(reichart_model)[[1]],
                                  MOFAcellulaR::project_data(model = reichart_model,
                                                  test_data = chaffin_multiview_dat))


# UMAP of factor
reichart_umap <- uwot::umap(reichart_factor_space, n_neighbors = 7)
colnames(reichart_umap) <- c("UMAP_1", "UMAP_2")

reichart_umap <- reichart_umap %>%
  as.data.frame() %>%
  rownames_to_column("sample") %>%
  left_join(all_meta, by = "sample") %>%
  dplyr::mutate(HF = ifelse(disease == "NF", "no", "yes"))

# First plot Chaffin space
reichart_space <- reichart_umap %>%
  dplyr::filter(study == "Reichart2022") %>%
  ggplot(., aes(x = UMAP_1,
                y = UMAP_2,
                color = HF)) +
  geom_point(size = 2, stroke = 1) +
  theme_classic() +
  theme(axis.text = element_text(size =12)) +
  scale_color_manual(values = c("grey", "black")) +
  xlim(c((reichart_umap$UMAP_1 %>% min()) - 1,
         (reichart_umap$UMAP_1 %>% max()) + 1)) +
  ylim(c((reichart_umap$UMAP_2 %>% min()) - 1,
         (reichart_umap$UMAP_2 %>% max()) + 1))

reichart_proj <- reichart_umap %>%
  dplyr::mutate(study = factor(study,
                               levels = c("Reichart2022", "Chaffin2022"))) %>%
  ggplot(., aes(x = UMAP_1,
                y = UMAP_2,
                color = HF,
                shape = study)) +
  geom_point(size = 2, stroke = 1) +
  theme_classic() +
  theme(axis.text = element_text(size =12)) +
  scale_color_manual(values = c("grey", "black")) +
  xlim(c((reichart_umap$UMAP_1 %>% min()) - 1,
         (reichart_umap$UMAP_1 %>% max()) + 1)) +
  ylim(c((reichart_umap$UMAP_2 %>% min()) - 1,
         (reichart_umap$UMAP_2 %>% max()) + 1))

panel_reichart <- cowplot::plot_grid(reichart_space, reichart_proj, align = "hv", ncol = 1)

pdf("./results/META/reichart_UMAP.pdf", height = 4.5, width = 3.8)
plot(panel_reichart)
dev.off()
```
